+++ src/main/java/duke/command/Command.java
@@ -1,4 +1,9 @@
-import duke.exceptions.DukeException;
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
