+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,4 +1,10 @@
-import duke.exceptions.DukeException;
+package duke.command;
+
+import duke.Storage;
+import duke.task.Task;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
 
 public class DeleteCommand extends Command {
