+++ src/main/java/duke/command/ExitCommand.java
@@ -1,4 +1,8 @@
-import duke.exceptions.DukeException;
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
 
 public class ExitCommand extends Command {
